pyproject.toml
[project]
name = "recipe-bot"
version = "0.0.1"
description = "Langgraph agent for personal recipe."
requires-python = "3.12"

dependencies = [
"python-dotenv==1.1.1",
"chroma==0.2.0",
"langchain==0.3.27",
"langchain-community==0.3.27",
"langchain-core==0.3.74",
"langchain-chroma==0.2.5",
"langchain-experimental==0.3.4",
"langchain-ollama==0.3.6",
"langchain-text-splitters==0.3.9",
"langchainhub==0.1.21",
"langgraph==0.6.2",
"langgraph-checkpoint==2.1.1",
"langgraph-sdk==0.2.0",
"langsmith==0.4.14",
"ollama==0.5.1",
"pydantic==2.11.7",
"pydantic-settings==2.10.1",
"pydantic_core==2.33.2",
"tavily_python==0.7.10",
"docling==2.43.0",
"pypdf==5.9.0",
"docx2txt==0.9",
"chromadb==1.0.17"
]

[project.optional-dependencies]
dev = [
    "ruff==0.4.10",
    "mypy==1.10.0",
    "pytest==8.2.2",
    "langgraph-cli==0.1.43",
]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
asyncio_mode = "auto"
